requirements.txt
Flask
PyPDF2
pyttsx3
uuid
python-docx
beautifulsoup4
requests